backend/pyproject.toml
[project]
name = "django-project-default"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.6.2,<5.7.0",
    "django>=5.2.0,<5.3.0",
    "djangorestframework>=3.16.0,<3.17",
    "djangorestframework-simplejwt>=5.5.0,<5.6.0",
    "drf-spectacular>=0.29.0,<0.30.0",
    "psycopg2>=2.9.0,<3.0.0",
    "python-dotenv>=1.2.0,<1.3.0",
    "redis>=7.2.1,<7.3.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "ruff>=0.15.4",
    "pytest>=9.0.2",
    "pytest-django>=4.12.0",
]

[tool.ruff]
line-length = 120
target-version = "py312"

exclude = [
    "migrations",
    "__pycache__",
    "manage.py",
    "settings.py",
    "tests/conftest.py"
]

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
]

[tool.ruff.lint.per-file-ignores]
"tests/conftest.py" = ["F403"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "app.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "-ra -q"
